Psalm Level 1 Testing: CostController, CostdetailController, CostheaderController
In order to prevent UndefinedMagicPropertyFetch and UndefinedMagicPropertyAssignment errors, getters and setters have been included in the models to prevent such errors.

Importantly to note for Psalm to work effectively:

````autoloader="vendor/yiisoft/yii2/Yii.php" ````

must be included in the Psalm.xml file in the root directory.

Full file as follows:

````
<?xml version="1.0"?>
<psalm
    autoloader="vendor/yiisoft/yii2/Yii.php"
    errorLevel="1"
    findUnusedBaselineEntry="true"
    findUnusedCode="false"
    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
    xmlns="https://getpsalm.org/schema/config"
    xsi:schemaLocation="https://getpsalm.org/schema/config vendor/vimeo/psalm/config.xsd"
>

<projectFiles>
    <directory name="frontend"/>
    <ignoreFiles>
        <directory name="vendor"/>
    </ignoreFiles>
</projectFiles>
</psalm>

```

````
<?php
declare(strict_types=1);
````

has been included at the top of all the models and conrtrollers.